navigationn to add device screen
