+++ app/src/main/kotlin/app/ui/Navigation.kt
@@ -14,9 +14,10 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
 import androidx.navigation.navDeepLink
-import app.activity.DevicesRoute
 import kotlinx.coroutines.flow.StateFlow
 import androidx.compose.runtime.CompositionLocalProvider
+import app.ui.screens.AddDevice
+import app.ui.screens.DevicesList
 
 sealed class Screen(val route: String) {
     object Devices : Screen("devices")
@@ -26,14 +27,9 @@ sealed class Screen(val route: String) {
 @Composable
 fun AppNavGraph(
     modifier: Modifier = Modifier,
-    deepLinkState: StateFlow<Intent?>,
     startDestination: String = Screen.Devices.route
 ) {
     val navController = rememberNavController()
-
-    deepLinkState.collectAsState(null).value?.let {
-        navController.handleDeepLink(it)
-    }
 
     NavHost(
         navController = navController,
@@ -49,10 +45,9 @@ internal fun NavGraphBuilder.addDevicesRoute(
     navController: NavHostController
 ) {
     composable(
-        route = Screen.Devices.route,
-        deepLinks = (getDeepLink("devices"))
+        route = Screen.Devices.route
     ) {
-        DevicesRoute(onAddDeviceClicked = {
+        DevicesList(onAddDeviceClicked = {
             navController.navigate(Screen.AddDevice.route)
         })
     }
@@ -62,23 +57,8 @@ internal fun NavGraphBuilder.addAddDeviceRoute(
     navController: NavHostController
 ) {
     composable(
-        route = Screen.Devices.route,
-        deepLinks = (getDeepLink("addDevice"))
+        route = Screen.AddDevice.route
     ) {
-        // TODO: Add AddDeviceRoute
-        //AddDeviceRoute(onBack = {
-//            navController.navigateUp()
-//        })
+        AddDevice(navigation = navController)
     }
-}
-
-fun getDeepLink(screen: String, arguments: String? = null): List<NavDeepLink> {
-    return listOf(
-        navDeepLink {
-            uriPattern = "batterybird://$screen" + if (arguments != null) "?$arguments" else ""
-        },
-        navDeepLink {
-            uriPattern = "https://batterybird.app/$screen" + if (arguments != null) "?$arguments" else ""
-        }
-    )
 }
