+++ app/src/main/kotlin/app/ui/widgets/TopAppBar.kt
@@ -1,0 +1,26 @@
+package app.ui.widgets
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavHostController
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun TopAppBar(title: String, navigation: NavHostController) {
+    TopAppBar(
+        title = { Text(title) },
+        navigationIcon = {
+            if (navigation.previousBackStackEntry != null) {
+                IconButton(onClick = { navigation.navigateUp() }) {
+                    Icon(Icons.Filled.ArrowBack, contentDescription = "Back")
+                }
+            }
+        }
+    )
+}
